pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# MyQuant MCP Server package
[project]
name = "myquant-mcp-server"
version = "0.1.0"
description = "MCP server for 掘金量化 (MyQuant) financial data and trading operations"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "MyQuant MCP", email = "noreply@example.com" }
]
license = { text = "MIT" }
keywords = ["mcp", "finance", "trading", "myquant", "gm-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "gm-python-sdk>=3.0.0",
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/myquant-mcp"
Documentation = "https://github.com/yourusername/myquant-mcp/blob/main/README.md"
Repository = "https://github.com/yourusername/myquant-mcp.git"
Issues = "https://github.com/yourusername/myquant-mcp/issues"

[project.scripts]
myquant-mcp = "windows_server:main"
